Минимальная реализация REST API с помощью аннотации @RepositoryRestResource (реализация с помощью класса BookController.java тоже работала, но удалена для краткости)
