+++ dz15-spring-rest-angular/spring-rest-backend/pom.xml
@@ -4,7 +4,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>org.springframework</groupId>
-    <artifactId>gs-accessing-mongodb-data-rest</artifactId>
+    <artifactId>dz15-spring-rest-angular</artifactId>
     <version>0.1.0</version>
 
     <parent>
