--- dz15-spring-rest-angular/spring-rest-backend/src/main/java/hello/Person.java
@@ -1,27 +1,0 @@
-package hello;
-
-import org.springframework.data.annotation.Id;
-
-public class Person {
-
-	@Id private String id;
-
-	private String firstName;
-	private String lastName;
-
-	public String getFirstName() {
-		return firstName;
-	}
-
-	public void setFirstName(String firstName) {
-		this.firstName = firstName;
-	}
-
-	public String getLastName() {
-		return lastName;
-	}
-
-	public void setLastName(String lastName) {
-		this.lastName = lastName;
-	}
-}
