--- dz15-spring-rest-angular/spring-rest-backend/src/main/java/hello/PersonRepository.java
@@ -1,15 +1,0 @@
-
-package hello;
-
-import java.util.List;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.data.repository.query.Param;
-import org.springframework.data.rest.core.annotation.RepositoryRestResource;
-
-@RepositoryRestResource(collectionResourceRel = "people", path = "people")
-public interface PersonRepository extends MongoRepository<Person, String> {
-
-	List<Person> findByLastName(@Param("name") String name);
-
-}
