+++ dz15-spring-rest-angular/spring-rest-backend/src/main/java/restapi/Application.java
@@ -1,5 +1,5 @@
 
-package hello;
+package restapi;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
