+++ dz15-spring-rest-angular/spring-rest-backend/src/main/java/restapi/BookRepository.java
@@ -1,4 +1,4 @@
-package hello;
+package restapi;
 
 import java.util.List;
 
@@ -7,10 +7,9 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import org.springframework.web.bind.annotation.CrossOrigin;
 
-// collectionResourceRel - не знаю что
+// collectionResourceRel - имя коллекции в mongodb
 // booklist - api endpoint (http://localhost:8080/booklist)
 @RepositoryRestResource(collectionResourceRel = "books", path = "booklist")
 @CrossOrigin(origins = "http://localhost:3000")
 public interface BookRepository extends MongoRepository<Book, String> {
-	//List<Person> findByLastName(@Param("name") String name);
 }
